refactored multi-array-expressions.h into expressionR.h and expressionRW.h

+++ include-templates/fun_unary_real_functions.h
@@ -12,7 +12,7 @@
 // ##COMMENTNAME## 
 //----------------------------------------------
 
-template <class X, class Element, typename Number, int depth, int rank, typename = std::enable_if_t<std::is_arithmetic<Number>::value>>
+template <class X, class Element, typename Number, size_t depth, size_t rank, typename = std::enable_if_t<std::is_arithmetic<Number>::value>>
 inline auto ##FUNCTION##(const ExpressionR<X, Element, ##DIN##, depth, rank>& x) {
   typedef typename NumberTrait<Element, ##DOUT##>::ReplacedNumberType EOUT;
   return  TER_Unary<ExpressionR<X, Element, ##DIN##, depth, rank>, EOUT, ##DOUT##, depth, rank, ##FUNCTOR##<Element, EOUT, ##DIN##, ##DOUT##>>(x);
